Convert MCPCore, ASTChunker, RAGCore, MCPServerKit to remote SPM deps
- Remove local package directories for MCPCore, ASTChunker, RAGCore, MCPServerKit
- Add remote SPM references to crunchybananas/{MCPCore,ast-chunker,RAGCore}
- Update Git/Package.swift MCPCore dep from path to URL
- Update ast-chunker-cli build script to clone from GitHub
- Add MCPCore as direct product dependency on macOS target
- Sync local ASTChunker improvements upstream (chunker enhancements)
- Push MCPCore agent prompt improvements upstream
- Tag MCPCore 1.1.0, ast-chunker 1.1.0, RAGCore 1.0.0
- Remaining local packages: Git, Github, Brew, PeelUI, PIIScrubber
- Also includes: RAG store refactor, PeonPing service, sound resources

+++ Shared/AgentOrchestration/Models/AgentChain.swift
@@ -481,9 +481,9 @@ public struct PrePlannerOutput: Sendable {
     public let errorSignature: String?
     public let fixDescription: String
     public let fixCode: String?
-    public let confidence: Float
+    public let confidence: Double
     
-    public init(id: String, filePattern: String?, errorSignature: String?, fixDescription: String, fixCode: String?, confidence: Float) {
+    public init(id: String, filePattern: String?, errorSignature: String?, fixDescription: String, fixCode: String?, confidence: Double) {
       self.id = id
       self.filePattern = filePattern
       self.errorSignature = errorSignature
